copy third-party libs instead of using Composer autoloader (#256) (#260)
The autoloader mechanism is quite heavy for a single dependency and
requires composer and vendor dir to be part of the final product.

We yet have generated/minified resources in our project, so this is the
minimalistic approach to copy the CrawlerDetect lib, too and extend the
custom autoloader mechanism.