fix text alignment in dashboard widget (#263) (#264)
We use RTL direction for the chart container in order see the rightmost
data when the chart has a horizontal scrollbar. This however results in
wrong text layout if there is no data or an error occurred.

Revert the text direction for all child elements and remove the custom
flex overrides for the chart labels.