do not return NULL, if database is empty (#273)
Null causes trouble during further processing, i.e. PHP warnings in the
API processing and later on problems building the dashboard widget in
Javascript, because all methods expect the array to have all keys.

Move the caching (transient) directive into the conditional, so we still
do not cache empty results and so not return NULL anymore.