move invisible tooltip to the top left corner to prevent layout shift
The tooltip is placed at the bottom of the body which makes the page
slightly longer, even though the user can't see the content. This might
result in a layout shift when the tooltip is displayed for the first
time and the scrollbar disappears.

Move the invisible tooltip element to the top-left corner fixes this
issue. We use left/top CSS properties, because these two will be
dynamically overwritten on mouse events anyway.