fix parameter types for target validation
The function wp_validate_redirect() expects a string as fallback
argument. This is not validated at runtime, so our boolean "false" is
passed through.

To comply with the API contracts, we now omit the second argument and
use an empty() check instead. We also continue with the output of the
validation call as its output is sanitized.